fixed type of s_top to handle signs

+++ lambton/2020/summer/ese2025/week_3/workspace/proj_stacks_no-pointers_offset/source/stacks_np.c
@@ -26,7 +26,7 @@
  ************************************************************************
  */
 int s[L]; // stack storage
-size_t s_top = -1;	// the stack "pointer", s.top, valid from 1 to L;
+long s_top = -1;	// the stack "pointer", s.top, valid from 1 to L;
 // a value of zero means the stack is empty
 
 /************************************************************************
